Improved note and get_date. Updated loggin format
