--- additional_functions/VA_date.py
@@ -1,65 +1,0 @@
-import datetime
-
-MONTHS = ['january', 'february', 'march', 'april', 'may', 'june',
-          'july', 'august', 'september', 'october', 'november', 'december']
-DAYS = ['monday', 'tuesday', 'wednesday', 'thursday', 'friday', 'saturday', 'sunday']
-DAY_EXTENSIONS = ["nd", "rd", "th", "st"]
-
-
-def get_date(text):
-    text = text.lower()
-    today = datetime.date.today()
-
-    if text.count("today") > 0:
-        return today
-
-    day = -1
-    day_of_week = -1
-    month = -1
-    year = today.year
-
-    tomorrow = datetime.date.today() + datetime.timedelta(days=1)
-    yesterday = datetime.date.today() - datetime.timedelta(days=1)
-    for word in text.split():
-        if word in MONTHS:
-            month = MONTHS.index(word) + 1
-        elif word in DAYS:
-            day_of_week = DAYS.index(word)
-        elif word.isdigit():
-            day = int(word)
-        else:
-            for ext in DAY_EXTENSIONS:
-                found = word.find(ext)
-                if found > 0:
-                    try:
-                        day = int(word[:found])
-                    except:
-                        pass
-
-    if month < today.month and month != -1:
-        year = year + 1
-
-    if day < today.day and month == -1 and day != -1:
-        month = month + 1
-
-    if month == -1 and day == -1 and day_of_week != -1:
-        current_day_of_week = today.weekday()
-        dif = day_of_week - current_day_of_week
-
-        if dif < 0:
-            dif += 7
-            if text.count("next") >= 1:
-                dif += 7
-
-        return today + datetime.timedelta(dif)
-
-    if text.count("tomorrow") > 0:
-        return tomorrow
-
-    if text.count("yesterday") > 0:
-        return yesterday
-
-    if month == -1 or day == -1:
-        return None
-
-    return datetime.date(month=month, day=day, year=year)
