--- additional_functions/VA_note.py
@@ -1,11 +1,0 @@
-import os
-import datetime
-import subprocess
-
-
-def note(text):
-    date = datetime.datetime.now()
-    file_name = str(date).replace(":", "-") + "-note.txt"
-    with open(os.path.join("../notes", file_name), "w") as f:
-        f.write(text)
-    subprocess.Popen(["notepad.exe", os.path.join("../notes", file_name)])
